requirements.txt
black==25.11.0
blinker==1.9.0
click==8.3.1
coverage==7.12.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gherkin-official==29.0.0
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
pytest-bdd==8.1.0
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dotenv==1.2.1
pytokens==0.3.0
ruff==0.14.7
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
WTForms==3.2.1
